v1.1.7
Feature added. 'cell_ref' argument now allows data to be imported to a specified cell. User request from Stein Rørvik (steinr on forum.image.sc). NOTE: the plugin does not check to see if data is already present at the specified location, so overwriting is possible. Column headers are also not exported with the data.